Final upload for FLS week 7
Homework - broken into 2 PDF versions of 2 separate PPT files.  These are the intended FLS to be turned in.  I've also printed the code from RStudio in a PDF.  There are 3 separate .Rmd files for this work.  In Part 2, my code stopped running so I have to build another .Rmd to execute the final problem, irris.